requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
apache-airflow==2.7.1
apache-airflow-providers-common-sql==1.7.2
apache-airflow-providers-ftp==3.5.2
apache-airflow-providers-http==4.5.2
apache-airflow-providers-imap==3.3.2
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-sqlite==3.4.3
apispec==6.3.0
argcomplete==3.1.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
backoff==2.2.1
blinker==1.6.2
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
constantly==23.10.4
contourpy==1.2.0
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.7
cssselect==1.2.0
cycler==0.12.1
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.3
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
google-re2==1.1
googleapis-common-protos==1.60.0
gradio==4.2.0
gradio_client==0.7.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.58.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.19.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.2
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
ordered-set==4.1.0
orjson==3.9.10
packaging==23.1
pandas==2.1.3
parsel==1.8.1
pathspec==0.11.2
pendulum==2.1.2
Pillow==10.1.0
pluggy==1.3.0
prison==0.2.1
Protego==0.3.0
protobuf==4.24.3
psutil==5.9.5
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.5.3
rich-argparse==1.3.0
rpds-py==0.10.3
Scrapy==2.11.0
semantic-version==2.10.0
service-identity==23.1.0
setproctitle==1.3.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
starlette==0.27.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
tldextract==5.1.1
tomlkit==0.12.0
toolz==0.12.0
tqdm==4.66.1
Twisted==22.10.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==2.0.4
uvicorn==0.24.0.post1
w3lib==2.1.2
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.17.0
zope.interface==6.1
